% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotowanie_danych.R
\name{.analizuj_skladnia_mcda}
\alias{.analizuj_skladnia_mcda}
\title{Wewnętrzny parser modelu badawczego}
\usage{
.analizuj_skladnia_mcda(skladnia)
}
\description{
Funkcja pomocnicza dekodująca definicje kryteriów ustalone przez badacza.
Przekształca zapis "Kryterium =~ pytanie1 + pytanie2" w strukturę listową.
}
\keyword{internal}
