% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver_metoda_BWM.R
\name{.buduj_rownanie_lp}
\alias{.buduj_rownanie_lp}
\title{Pomocnik budowania ograniczeń dla solvera LP}
\usage{
.buduj_rownanie_lp(macierz_ograniczen, nowe_rownanie)
}
\description{
Pomocnik budowania ograniczeń dla solvera LP
}
\keyword{internal}
