% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rozmyte_mcda.R
\name{.oblicz_dominacje_erasmus}
\alias{.oblicz_dominacje_erasmus}
\title{Wewnętrzna Funkcja Dominacji dla Meta-Rankingu}
\usage{
.oblicz_dominacje_erasmus(r1, r2, r3)
}
\arguments{
\item{r1}{Wektor numeryczny rang metody TOPSIS.}

\item{r2}{Wektor numeryczny rang metody VIKOR.}

\item{r3}{Wektor numeryczny rang metody WASPAS.}
}
\value{
Wektor numeryczny z finalnym rankingiem.
}
\description{
Funkcja pomocnicza wyznaczająca ranking konsensusu na podstawie reguły większości.
Rozstrzyga, która uczelnia najczęściej wygrywa na danej pozycji w trzech różnych algorytmach.
}
\keyword{internal}
