% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotowanie_danych.R
\name{.skaluj_do_saaty}
\alias{.skaluj_do_saaty}
\title{Wewnętrzny Skaler Saaty'ego - funkcja wewnętrzna 2}
\usage{
.skaluj_do_saaty(odpowiedzi_studentow)
}
\description{
Przekształca dowolną skalę (np. Likert 1-5, wartości ciągłe)
na skalę Saaty'ego 1-9.
}
\keyword{internal}
