% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotowanie_danych.R
\name{.transformuj_na_tfn}
\alias{.transformuj_na_tfn}
\title{Wewnętrzny Fuzzifier wyników uczelni}
\usage{
.transformuj_na_tfn(srednie_ostre)
}
\description{
Wprowadza niepewność ocen studenckich poprzez konwersję tzw. oceny ostrej (Crisp) 
na Trójkątną Liczbę Rozmytą (TFN - Triangular Fuzzy Number).
}
\keyword{internal}
