% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rozmyte_mcda.R
\name{.ustal_wagi_kryteriow_mobilnosci}
\alias{.ustal_wagi_kryteriow_mobilnosci}
\title{Wewnętrzny menedżer priorytetów ankiet}
\usage{
.ustal_wagi_kryteriow_mobilnosci(
  rozmyta_macierz_decyzyjna,
  wagi_odgorne,
  nazwy_kryteriow_bwm,
  ankieta_najwazniejszy_vs_reszta,
  ankieta_reszta_vs_najmniej_wazny
)
}
\description{
Decyduje, skąd pobrać wagi dla kryteriów wyjazdu (np. ustalone odgórnie vs z modelu BWM).
}
\keyword{internal}
