% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_worst_method_solver.R
\name{.waliduj_priorytety_studenta}
\alias{.waliduj_priorytety_studenta}
\title{Wewnętrzna walidacja ankiety BWM}
\usage{
.waliduj_priorytety_studenta(
  najwazniejszy_vs_reszta,
  reszta_vs_najmniej_wazny,
  nazwy_kryteriow
)
}
\description{
Sprawdza, czy wektory porównań od ankietowanych spełniają zasady Best-Worst.
}
\keyword{internal}
