% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_worst_method_solver.R
\name{.wymus_logike_bwm}
\alias{.wymus_logike_bwm}
\title{Wewnętrzne asercje logiki BWM}
\usage{
.wymus_logike_bwm(wyrazenie, komunikat)
}
\description{
Sprawdza, czy preferencje studentów (wagi kryteriow) mają sens matematyczny.
}
\keyword{internal}
