% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rozmyte_mcda.R
\name{wyznacz_meta_ranking_erasmus}
\alias{wyznacz_meta_ranking_erasmus}
\title{Ostateczny Meta-Ranking Uczelni (Konsensus z trzech metod)}
\usage{
wyznacz_meta_ranking_erasmus(
  rozmyta_macierz_decyzyjna,
  kierunki_kryteriow,
  wagi_odgorne = NULL,
  nazwy_kryteriow_bwm = NULL,
  bwm_najlepsze = NULL,
  bwm_najgorsze = NULL,
  wspolczynnik_kompromisu_v = 0.5,
  wspolczynnik_lambda = 0.5
)
}
\arguments{
\item{rozmyta_macierz_decyzyjna}{Rozmyta macierz danych ze wskaźnikami uczelni.}

\item{kierunki_kryteriow}{Wektor kierunków optymalizacji ("min", "max").}

\item{wagi_odgorne}{(Opcjonalnie) Wagi kryteriów wyliczone np. metodą Entropii.}

\item{nazwy_kryteriow_bwm}{(Opcjonalnie) Nazwy kryteriów do modelu BWM.}

\item{bwm_najlepsze}{(Opcjonalnie) Wektor priorytetów Best-to-Others.}

\item{bwm_najgorsze}{(Opcjonalnie) Wektor priorytetów Others-to-Worst.}

\item{wspolczynnik_kompromisu_v}{Parametr dla VIKOR (domyślnie 0.5).}

\item{wspolczynnik_lambda}{Parametr dla WASPAS (domyślnie 0.5).}
}
\value{
Lista zawierająca ramkę danych z zestawieniem wszystkich rankingów oraz macierz korelacji.
}
\description{
Agreguje wyniki z metod Fuzzy TOPSIS, Fuzzy VIKOR i Fuzzy WASPAS, aby stworzyć
jeden, absolutnie odporny ranking końcowy dla mobilności studenckiej.
}
