% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rozmyte_mcda.R
\name{wyznacz_ranking_waspas}
\alias{wyznacz_ranking_waspas}
\title{Rozmyta Metoda WASPAS dla Wyboru Uczelni}
\usage{
wyznacz_ranking_waspas(
  rozmyta_macierz_decyzyjna,
  kierunki_kryteriow,
  wspolczynnik_lambda = 0.5,
  wagi_odgorne = NULL,
  nazwy_kryteriow_bwm = NULL,
  bwm_najlepsze = NULL,
  bwm_najgorsze = NULL
)
}
\arguments{
\item{rozmyta_macierz_decyzyjna}{Macierz m x 3n wygenerowana ze wskaźników uczelni.}

\item{kierunki_kryteriow}{Wektor znakowy ("max" dla kryteriów pozytywnych, "min" dla negatywnych).}

\item{wspolczynnik_lambda}{Parametr równowagi między WSM a WPM (domyślnie 0.5 oznacza równy wpływ).}

\item{wagi_odgorne}{(Opcjonalnie) Gotowy wektor wag dla kryteriów.}

\item{nazwy_kryteriow_bwm}{(Opcjonalnie) Nazwy kryteriów do modelu BWM.}

\item{bwm_najlepsze}{(Opcjonalnie) Wektor Best-to-Others.}

\item{bwm_najgorsze}{(Opcjonalnie) Wektor Others-to-Worst.}
}
\value{
Obiekt klasy `waspas_erasmus_wynik` z rankingiem uczelni.
}
\description{
Implementacja kompromisowej metody Fuzzy WASPAS w kontekście mobilności studenckiej.
Łączy dwa skrajne podejścia matematyczne: addytywne (Suma Ważona - WSM) 
oraz multiplikatywne (Iloczyn Ważony - WPM), co znacznie zwiększa dokładność 
i stabilność ostatecznego rankingu.
}
