% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rozmyte_mcda.R
\name{wyznacz_wagi_obiektywne_entropia}
\alias{wyznacz_wagi_obiektywne_entropia}
\title{Obliczanie obiektywnych wag metodą Entropii Shannona}
\usage{
wyznacz_wagi_obiektywne_entropia(rozmyta_macierz_decyzyjna)
}
\arguments{
\item{rozmyta_macierz_decyzyjna}{Rozmyta macierz uczelni (wynik funkcji `zbuduj_macierz_rozmyta`).}
}
\value{
Wektor numeryczny wag kryteriów sumujący się do 1 (100%).
}
\description{
Wyznacza obiektywne wagi kryteriów na podstawie danych z ankiet studenckich,
mierząc stopień rozproszenia ocen. Im większa zmienność opinii o danym kryterium, tym wyższa jego waga.
}
