% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver_metoda_BWM.R
\name{.ocen_spojnosc_odpowiedzi}
\alias{.ocen_spojnosc_odpowiedzi}
\title{Wewnętrzna ocena spójności wywiadów}
\usage{
.ocen_spojnosc_odpowiedzi(model_badawczy)
}
\description{
Wewnętrzna ocena spójności wywiadów
}
\keyword{internal}
