% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wizualizacja.R
\name{plot.vikor_erasmus_wynik}
\alias{plot.vikor_erasmus_wynik}
\title{Mapa Strategiczna VIKOR}
\usage{
\method{plot}{vikor_erasmus_wynik}(x, ...)
}
\arguments{
\item{x}{Obiekt klasy `rozmyty_vikor_wynik`.}

\item{...}{Dodatkowe argumenty (ignorowane).}
}
\description{
Wizualizacja typu cIPMA.
Oś X: Efektywność grupowa (odwrócone S). Oś Y: Ryzyko/Żal (R).
Wielkość bąbla: Siła kompromisu (zależna od Q).
}
