% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wizualizacja.R
\name{plot.waspas_erasmus_wynik}
\alias{plot.waspas_erasmus_wynik}
\title{Mapa Balansu WASPAS}
\usage{
\method{plot}{waspas_erasmus_wynik}(x, ...)
}
\arguments{
\item{x}{Obiekt klasy `waspas_erasmus_wynik`.}

\item{...}{Dodatkowe argumenty.}
}
\description{
Pokazuje relację między modelem sumarycznym (WSM) a iloczynowym (WPM).
Oś X: Suma Ważona (WSM). Oś Y: Iloczyn Ważony (WPM).
Uczelnie w prawym górnym rogu są najlepsze w obu modelach.
}
