% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rozmyte_mcda.R
\name{wyznacz_ranking_topsis}
\alias{wyznacz_ranking_topsis}
\title{Rozmyta Metoda TOPSIS dla Wyboru Uczelni}
\usage{
wyznacz_ranking_topsis(
  rozmyta_macierz_decyzyjna,
  kierunki_kryteriow,
  wagi_odgorne = NULL,
  nazwy_kryteriow_bwm,
  bwm_najlepsze,
  bwm_najgorsze
)
}
\arguments{
\item{rozmyta_macierz_decyzyjna}{Macierz m x 3n.}

\item{kierunki_kryteriow}{Wektor znakowy ("max" dla kryteriów pozytywnych np. stypendium, "min" dla negatywnych np. koszty).}

\item{wagi_odgorne}{(Opcjonalnie) Gotowy wektor wag dla kryteriów (np. z metody Entropii).}

\item{nazwy_kryteriow_bwm}{(Opcjonalnie) Nazwy kryteriów do dynamicznego wyliczenia wag z ankiet BWM.}

\item{bwm_najlepsze}{(Opcjonalnie) Wektor Best-to-Others.}

\item{bwm_najgorsze}{(Opcjonalnie) Wektor Others-to-Worst.}
}
\value{
Obiekt klasy `topsis_erasmus_wynik` z rankingiem uczelni wyjazdowych.
}
\description{
Implementacja algorytmu Fuzzy TOPSIS w kontekście mobilności studenckiej.
Oblicza odległość każdej badanej uczelni od wyidealizowanego rozwiązania perfekcyjnego (FPIS)
oraz najgorszego rozwiązania anty-idealnego (FNIS).
}
