% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rozmyte_mcda.R
\name{wyznacz_ranking_vikor}
\alias{wyznacz_ranking_vikor}
\title{Rozmyta Metoda VIKOR dla Wyboru Uczelni}
\usage{
wyznacz_ranking_vikor(
  rozmyta_macierz_decyzyjna,
  kierunki_kryteriow,
  wspolczynnik_kompromisu_v = 0.5,
  wagi_odgorne = NULL,
  nazwy_kryteriow_bwm,
  bwm_najlepsze,
  bwm_najgorsze
)
}
\arguments{
\item{rozmyta_macierz_decyzyjna}{Macierz ($m \times 3n$) wygenerowana ze wskaźników uczelni.}

\item{kierunki_kryteriow}{Wektor znakowy ("max" dla kryteriów pozytywnych, "min" dla negatywnych).}

\item{wspolczynnik_kompromisu_v}{Waga strategii "większości kryteriów" (domyślnie 0.5 oznacza równowagę).}

\item{wagi_odgorne}{(Opcjonalnie) Gotowy wektor wag dla kryteriów.}

\item{nazwy_kryteriow_bwm}{(Opcjonalnie) Nazwy kryteriów do dynamicznego wyliczenia wag z ankiet BWM.}

\item{bwm_najlepsze}{(Opcjonalnie) Wektor Best-to-Others.}

\item{bwm_najgorsze}{(Opcjonalnie) Wektor Others-to-Worst.}
}
\value{
Obiekt klasy `vikor_erasmus_wynik` z rankingiem uczelni.
}
\description{
Implementacja kompromisowej metody Fuzzy VIKOR. Oblicza wskaźniki S (maksymalna użyteczność grupowa),
R (indywidualny żal decydenta) oraz Q (ostateczny indeks kompromisu) dla uczelni partnerskich.
}
