% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_worst_method_solver.R
\name{wyznacz_wagi_kryteriow_bwm}
\alias{wyznacz_wagi_kryteriow_bwm}
\title{Wyznaczanie Wag Kryteriów Metodą BWM (Erasmus+)}
\usage{
wyznacz_wagi_kryteriow_bwm(
  nazwy_kryteriow,
  najwazniejszy_vs_reszta,
  reszta_vs_najmniej_wazny
)
}
\arguments{
\item{nazwy_kryteriow}{Wektor znakowy z nazwami kryteriów (np. c("Finanse", "Jakosc")).}

\item{najwazniejszy_vs_reszta}{Wektor numeryczny (1-9). Przewaga NAJWAŻNIEJSZEGO kryterium nad pozostałymi.}

\item{reszta_vs_najmniej_wazny}{Wektor numeryczny (1-9). Przewaga pozostałych kryteriów nad NAJMNIEJ WAŻNYM.}
}
\value{
Lista zawierająca: `wagi_kryteriow`, `wskaznik_spojnosci` (CR) oraz wartość błędu `ksi`.
}
\description{
Oblicza optymalne wagi ważności dla głównych kryteriów oceny uczelni (np. Finanse, Jakość)
metodą Best-Worst (BWM) przy użyciu programowania liniowego. Minimalizuje błąd niespójności decydenta (ksi).
}
