% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotowanie_danych.R
\name{zbuduj_macierz_rozmyta}
\alias{zbuduj_macierz_rozmyta}
\title{funkcja publiczna
Przygotowanie Danych do Rozmytej Analizy MCDA}
\usage{
zbuduj_macierz_rozmyta(
  dane,
  skladnia,
  kolumna_uczelni = NULL,
  agregacja_studenci = mean
)
}
\arguments{
\item{dane}{Ramka danych (data frame) zawierająca surowe zmienne.}

\item{skladnia}{Ciąg znaków definiujący kryteria (np. "Koszt =~ k1 + k2").}

\item{kolumna_uczelni}{Nazwa kolumny identyfikującej alternatywy.
Jeśli NULL, każdy wiersz traktowany jest jako osobna alternatywa.}

\item{agregacja_studenci}{Funkcja używana do scalania opinii studentów (domyślnie: mean).}
}
\value{
Macierz o wymiarach m x 3n, gdzie m to liczba uczelni.
}
\description{
Funkcja przekształca surowe dane ankietowe w rozmytą macierz decyzyjną.
Oblicza wyniki zmiennych kompozytowych na podstawie składni, skaluje je do przedziału 1-9,
agreguje odpowiedzi ekspertów (jeśli dotyczy) i dokonuje rozmycia (fuzzification).
}
